requirements.txt
environs==9.2.0
python-telegram-bot==13.0
pymongo==3.11.0
requests==2.25.0
googletrans==3.0.0